The build on power6 was warning in a DEBUG statement because sizeof() returns an int rather than a long.  So use a cast to avoid this.

+++ src/papi_libpfm4_events.c
@@ -342,7 +342,7 @@ static struct native_event_t *allocate_native_event(char *name,
 
      SUBDBG("Allocating more room for native events (%d %ld)\n",
 	    (allocated_native_events+NATIVE_EVENT_CHUNK),
-	    sizeof(struct native_event_t) *
+	    (long)sizeof(struct native_event_t) *
 	    (allocated_native_events+NATIVE_EVENT_CHUNK));
 
      native_events=realloc(native_events,
